requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
DateTime==5.5
fastapi==0.115.7
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mysql-connector-python==9.2.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zope.interface==7.2
